refactor css files into Styles folder
