+++ src/pages/watchLaterList.js
@@ -1,7 +1,7 @@
 import {useVideoPlayerContext} from "../contexts/videoListcontext";
 import  ReactPlayer from "react-player";
-import "../displaycomponents/category/category.css";
-import "../displaycomponents/category/button.css";
+import "../styles/button.css";
+import "../styles/category.css";
 
 function WatchLaterVideoList(){
